Add containers/image/transports subpackage
This package maintains the list of all known transports, and provides
helpers to convert between strings and types.ImageReference ; this
should be the primary entrypoint for callers who want to just provide an
UI without dealing with the underlying ImageReference structures or
specific transports.

This makes the skopeo/cmd/skopeo/utils.go transport:transport-specific
string format a containers/image -defined feature.

This would fit nicely into containers/image/image, but the individual
transports depend on that package for creating a types.Image instance,
so we would have a circular dependency.  And containers/image/signature
will need to depend on the list of transports maintained by this, so
maintaining it at the top containers/image level would be a bad fit as
well.

So, sadly, another small subpackage.

Though perhaps it should be called /reference instead of /transports?
Either way, one half of the API will feel a bit off.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>