Do not duplicate OCI subdirectory paths
Instead of supplying literals to ensureParentDirectoryExists, first
compute the path of the file within the directory using the ociReference
methods, and then use that to compute the parent directory path.

Should not change behavior, except that filepath.Dir() implicitly calls
filepath.Clean(), apparently incompatible with UNIX path semantics.  But
then we always call filepath.Join() with the same d.ref.dir input, which
makes the same implicit call, and Go's handling of symlinks and .. seems
close-to-unfixably broken in general, and trying to fix that only here
is rather pointless:
https://github.com/golang/go/issues/4382

Signed-off-by: Miloslav Trmač <mitr@redhat.com>